update fake data function

+++ basic/migrations/m231017_065804_create_table.php
@@ -17,7 +17,7 @@ class m231017_065804_create_table extends Migration
             'password' => $this->string()->notNull(),
             'name' => $this->string()->notNull(),
             'last_name' => $this->string()->notNull(),
-            'phone' => $this->integer()->notNull(),
+            'phone' => $this->bigInteger()->notNull(),
             'admin' => $this->boolean()->defaultValue(0),
             'access_token' => $this->string()
         ]);
